moving audio instance to audio folder
